Update actions.js
I was working with a hog and when I tried to run a scene that was over 1000, the text turned red. The hog supports up to 32768 scenes, so I figured that I would PR so it does not turn red with a value that will still work with the console.